updating @After methods
