+++ activiti-spring-boot-starter/src/test/java/org/activiti/spring/boot/tasks/TaskRuntimeUnAuthorizedTest.java
@@ -8,6 +8,7 @@ import org.activiti.api.task.model.builders.TaskPayloadBuilder;
 import org.activiti.api.task.runtime.TaskAdminRuntime;
 import org.activiti.api.task.runtime.TaskRuntime;
 import org.activiti.spring.boot.security.util.SecurityUtil;
+import org.activiti.spring.boot.test.util.TaskCleanUpUtil;
 import org.junit.After;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
@@ -35,22 +36,16 @@ public class TaskRuntimeUnAuthorizedTest {
     @Autowired
     private SecurityUtil securityUtil;
 
-    @After
-    public void cleanUpWithAdmin() {
-        securityUtil.logInAs("admin");
-        Page<Task> tasks = taskAdminRuntime.tasks(Pageable.of(0, 50));
-        for (Task t : tasks.getContent()) {
-            taskAdminRuntime.delete(TaskPayloadBuilder
-                    .delete()
-                    .withTaskId(t.getId())
-                    .withReason("test clean up")
-                    .build());
-        }
+    @Autowired
+    private TaskCleanUpUtil taskCleanUpUtil;
 
+    @After
+    public void taskCleanUp(){
+        taskCleanUpUtil.cleanUpWithAdmin();
     }
 
-
-    public void aCreateStandaloneTaskForGroup() {
+    @Test
+    public void createStandaloneTaskForGroup() {
 
         securityUtil.logInAs("garth");
